requirements.txt
streamlit
torch
transformers
datasets
faiss-cpu
scikit-learn
python-docx
sentence-transformers
accelerate
peft
langchain
langchain-groq
python-dotenv